model HistoryDAO
